implemented the new analytics for post model
