--- app/Events/PostViewed.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace App\Events;
-
-use App\Model\Blog\Post;
-
-class PostViewed
-{
-  /**
-   * The post instance.
-   *
-   * @var Post
-   */
-  public $post;
-
-  /**
-   * Create a new event instance.
-   *
-   * @param Post $post
-   */
-  public function __construct(Post $post)
-  {
-    $this->post = $post;
-  }
-}
