--- resources/js/routes/admin/stats.js
@@ -1,12 +1,0 @@
-export default [
-  {
-    path: '/admin/stats',
-    name: 'stats',
-    component: require('../../screens/admin/stats/StatsIndex').default,
-  },
-  {
-    path: '/admin/stats/:id',
-    name: 'stats-show',
-    component: require('../../screens/admin/stats/StatsShow').default,
-  },
-]
